api/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
idna==3.6
iniconfig==2.0.0
model-bakery==1.17.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.1.1
pytest-django==4.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==69.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
urllib3==2.2.1
